refactor(fx-core): stop writing the legacy template-name onto v4 create inputs
The v4->v3 template id map is a generate-template telemetry key (DCE-21 OKR continuity), not a dispatch input. It moves next to its only consumer in createFrontDoorAdapters and derives from target.templateId, so the v4 front door no longer mutates the shared v3 inputs bag. Telemetry values are unchanged.
